requirements.txt
biopython==1.79
pandas==1.5.1
tqdm==4.64.1